Eliminate most GCC warnings (initializer ordering, signed/unsigned mismatches), but leaving auto_ptr and unused variable warnings alone
[git-p4: depot-paths = "//depot/openwatcom/": change = 35156]

+++ bld/wipfc/cpp/gdword.cpp
@@ -53,7 +53,7 @@ size_t GlobalDictionaryWord::writeWord( std::FILE* out ) const
     char buffer[ 256 ];
     size_t written;
     size_t length( std::wcstombs( buffer, text.c_str(), sizeof( buffer ) / sizeof( char ) ) );
-    if( length == -1 )
+    if( length == static_cast< size_t >( -1 ) )
         throw FatalError( ERR_T_CONV );
     if( length > 254 )
         length = 254;
